feat: manejo de codigo http, validacion de isSystem en servicios de accesos para restringir acciones

+++ src/services/accesso.service.ts
@@ -1,6 +1,7 @@
 import { NextFunction } from 'express';
 import { AccessDTO } from 'src/dto/access.dto';
 import HttpException from '../common/http.exception';
+import { HttpStatus } from '../constants/http-status';
 import * as accessRepository from '../repositories/accesso.repository';
 import { Messages } from './../constants/messages';
 
@@ -11,7 +12,12 @@ const getAllAccess = async function (
         const access = await accessRepository.getAllAccess();
         return access.map(mapAccess);
     } catch (error) {
-        next(new HttpException(500, Messages.GET_ACCESS_ERROR));
+        next(
+            new HttpException(
+                HttpStatus.INTERNAL_SERVER_ERROR,
+                Messages.GET_ACCESS_ERROR
+            )
+        );
     }
 };
 
@@ -20,11 +26,20 @@ const addAccess = async function (
     next: NextFunction
 ): Promise<AccessDTO | void> {
     try {
+        if (await checkIfExistAccessName(contract.name)) {
+            next(new HttpException(HttpStatus.CONFLICT, Messages.ACCESS_EXIST));
+            return;
+        }
         const data = await accessRepository.addAccess(contract);
         contract.id = data._id;
         return contract;
     } catch (error) {
-        next(new HttpException(500, Messages.ADD_ACCESS_ERROR));
+        next(
+            new HttpException(
+                HttpStatus.INTERNAL_SERVER_ERROR,
+                Messages.ADD_ACCESS_ERROR
+            )
+        );
     }
 };
 
@@ -33,27 +48,59 @@ const updateAccess = async function (
     next: NextFunction
 ): Promise<AccessDTO | void> {
     try {
+        if (
+            await accessRepository.checkIfExistAccessName(
+                contract.name,
+                contract.id
+            )
+        ) {
+            next(new HttpException(HttpStatus.CONFLICT, Messages.ACCESS_EXIST));
+            return;
+        }
         const data = await accessRepository.updateAccess(contract);
         if (!data) {
-            next(new HttpException(404, Messages.ACCESS_NOT_FOUND));
+            next(
+                new HttpException(
+                    HttpStatus.NOT_FOUND,
+                    Messages.ACCESS_NOT_FOUND
+                )
+            );
             return;
         }
         contract.id = data._id;
         return contract;
     } catch (error) {
-        next(new HttpException(404, Messages.ACCESS_NOT_FOUND));
+        next(
+            new HttpException(HttpStatus.NOT_FOUND, Messages.ACCESS_NOT_FOUND)
+        );
     }
 };
+
+function checkIfExistAccessName(name: string, id?: string) {
+    return accessRepository.checkIfExistAccessName(name, id);
+}
 
 const deleteAccess = async function (
     id: string,
     next: NextFunction
 ): Promise<boolean | void> {
     try {
+        const isSystem = await accessRepository.isSystem(id);
+        if (isSystem) {
+            next(
+                new HttpException(
+                    HttpStatus.CONFLICT,
+                    Messages.ACCESS_NOT_REMOVABLE
+                )
+            );
+            return;
+        }
         await accessRepository.deleteAccess(id);
         return true;
     } catch (error) {
-        next(new HttpException(404, Messages.ACCESS_NOT_FOUND));
+        next(
+            new HttpException(HttpStatus.NOT_FOUND, Messages.ACCESS_NOT_FOUND)
+        );
     }
 };
 
@@ -65,7 +112,9 @@ const findAccessById = async function (
         const data = await accessRepository.findById(id);
         return mapAccess(data);
     } catch (error) {
-        next(new HttpException(404, Messages.ACCESS_NOT_FOUND));
+        next(
+            new HttpException(HttpStatus.NOT_FOUND, Messages.ACCESS_NOT_FOUND)
+        );
     }
 };
 
@@ -74,13 +123,33 @@ const deleteAccessesByIds = async function (
     next: NextFunction
 ): Promise<boolean | void> {
     try {
+        if (await checkIfAnyIsFromSystem(ids)) {
+            next(
+                new HttpException(
+                    HttpStatus.CONFLICT,
+                    Messages.ACCESS_NOT_REMOVABLE
+                )
+            );
+            return;
+        }
         await accessRepository.deleteByIds(ids);
         return true;
     } catch (error) {
         console.log(error);
-        next(new HttpException(404, Messages.ACCESS_NOT_FOUND));
+        next(
+            new HttpException(HttpStatus.NOT_FOUND, Messages.ACCESS_NOT_FOUND)
+        );
     }
 };
+
+async function checkIfAnyIsFromSystem(ids: string[]): Promise<boolean> {
+    for (let index = 0; index < ids.length; index++) {
+        if (await accessRepository.isSystem(ids[index])) {
+            return true;
+        }
+    }
+    return false;
+}
 
 function mapAccess(item: any): AccessDTO {
     return {
@@ -88,7 +157,8 @@ function mapAccess(item: any): AccessDTO {
         name: item.name,
         route: item.route,
         description: item.description,
-        active: item.active
+        active: item.active,
+        isSystem: item.isSystem
     };
 }
